Refactor tokens and contracts
