+++ .env.example
@@ -60,13 +60,11 @@ TEST_ON_HARDHAT_NODE=
 # used to submit contract verifications to etherscan during hardhat deploys
 #ETHERSCAN_API_KEY=your-api-key
 
-# Public FE keys for Alchemy
-VITE_ALCHEMY_OPTIMISM_SEPOLIA_API_KEY=get-a-key-from-alchemy-opt-sepolia
-VITE_ALCHEMY_OPTIMISM_API_KEY=get-a-key-from-alchemy-opt-mainnet
-VITE_ALCHEMY_ETH_MAINNET_API_KEY=get-a-key-from-alchemy-eth-mainnet
-
+# Public FE key for Alchemy
+VITE_FE_ALCHEMY_API_KEY=get-a-key-from-alchemy
 
 #Backend private keys for Alchemy
+BE_ALCHEMY_API_KEY=get-a-key-from-alchemy
 
 # HASURA_GRAPHQL auto used by hasura CLI (used with ./hasura/config.yaml)
 HASURA_GRAPHQL_ENDPOINT=http://localhost:8080
