+++ src/utils/provider.ts
@@ -2,10 +2,6 @@ import type { Web3Provider } from '@ethersproject/providers';
 import { ethers } from 'ethers';
 import _ from 'lodash-es';
 
-import {
-  VITE_ALCHEMY_OPTIMISM_API_KEY,
-  VITE_ALCHEMY_OPTIMISM_SEPOLIA_API_KEY,
-} from '../config/env';
 import { chain } from '../features/cosoul/chains';
 
 export const getSignature = async (
@@ -145,21 +141,3 @@ export async function switchOrAddNetwork(
     }
   }
 }
-
-export const getReadOnlyProvider = (chainId: number) => {
-  switch (chainId) {
-    case 10:
-      return new ethers.providers.AlchemyProvider(
-        chainId,
-        VITE_ALCHEMY_OPTIMISM_API_KEY
-      );
-    case 11155420:
-      // ethers v6 required for optimism-sepolia support
-      return new ethers.providers.JsonRpcProvider(
-        `https://opt-sepolia.g.alchemy.com/v2/${VITE_ALCHEMY_OPTIMISM_SEPOLIA_API_KEY}`
-      );
-    case 1338:
-      // TODO: this is hacked for local - does it need to handle CI port?
-      return new ethers.providers.JsonRpcProvider(`http://localhost:8546`);
-  }
-};
