+++ src/utils/viem/publicClient.ts
@@ -3,12 +3,14 @@ import { localhost, optimism, optimismSepolia } from 'viem/chains';
 
 import {
   HARDHAT_GANACHE_PORT,
-  VITE_ALCHEMY_OPTIMISM_API_KEY,
-  VITE_ALCHEMY_OPTIMISM_SEPOLIA_API_KEY,
+  VITE_FE_ALCHEMY_API_KEY,
 } from '../../config/env';
 import { chain } from '../../features/cosoul/chains';
 
-export function getReadOnlyClient(chainId?: number) {
+export type ReadOnlyClient = ReturnType<typeof getReadOnlyClient>;
+export function getReadOnlyClient(chainId?: number, alchemyApiKey?: string) {
+  const apiKey = alchemyApiKey ?? VITE_FE_ALCHEMY_API_KEY;
+
   if (chainId === undefined) {
     chainId = Number(chain.chainId);
   }
@@ -17,16 +19,12 @@ export function getReadOnlyClient(chainId?: number) {
     case 10: // Optimism
       return createPublicClient({
         chain: optimism,
-        transport: http(
-          `https://opt-mainnet.g.alchemy.com/v2/${VITE_ALCHEMY_OPTIMISM_API_KEY}`
-        ),
+        transport: http(`https://opt-mainnet.g.alchemy.com/v2/${apiKey}`),
       });
     case 11155420: // Optimism Sepolia
       return createPublicClient({
         chain: optimismSepolia,
-        transport: http(
-          `https://opt-sepolia.g.alchemy.com/v2/${VITE_ALCHEMY_OPTIMISM_SEPOLIA_API_KEY}`
-        ),
+        transport: http(`https://opt-sepolia.g.alchemy.com/v2/${apiKey}`),
       });
     case 1338: // Local development/CI chain
       return createPublicClient({
